# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  server.apiErrorResponse:
    properties:
      error:
        type: string
    type: object
  server.apiResponse:
    properties:
      data: {}
      meta: {}
    type: object
...
  server.createMerchantRequest:
    properties:
      bpjs_ppk_code:
        type: string
      company_id:
        type: string
      kemkes_facility_code:
        type: string
      name:
        type: string
      timezone:
        type: string
    required:
    - company_id
    - name
    - timezone
    type: object
...
  server.updateCompanyRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
...
  server.updateMerchantRequest:
    properties:
      bpjs_ppk_code:
        type: string
      is_active:
        type: boolean
      kemkes_facility_code:
        type: string
      name:
        type: string
      timezone:
        type: string
    required:
    - name
    - timezone
    type: object
...
info:
  contact: {}
  description: 'Backend service for the NEXQIA Healthcare Operating Platform. All
    routes are under /api/v1. Response envelope: {"data": ..., "meta": {...}}.'
  title: NEXQIA API
  version: "1.0"
paths:
...
  /companies/{id}:
    delete:
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete a company
      tags:
      - tenancy
    get:
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a company by id
      tags:
      - tenancy
    patch:
      consumes:
      - application/json
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      - description: Company data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.updateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a company
      tags:
      - tenancy
...
  /companies/{id}/merchants:
    get:
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
      security:
      - BearerAuth: []
      summary: List merchants under a company
      tags:
      - tenancy
...
  /merchants:
    post:
      consumes:
      - application/json
      parameters:
      - description: Merchant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.createMerchantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.apiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a merchant
      tags:
      - tenancy
  /merchants/{id}:
    delete:
      parameters:
      - description: Merchant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Soft-delete a merchant
      tags:
      - tenancy
    get:
      parameters:
      - description: Merchant UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a merchant by id
      tags:
      - tenancy
    patch:
      consumes:
      - application/json
      parameters:
      - description: Merchant UUID
        in: path
        name: id
        required: true
        type: string
      - description: Merchant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.updateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a merchant
      tags:
      - tenancy
...
swagger: "2.0"